+++ Adventure1/MainMenuState.cpp
@@ -17,7 +17,7 @@ void MainMenuState::InitVal()
 
 void MainMenuState::InitBackground()
 {
-	this->background.setSize(Vector2f(this->window->getSize().x, this->window->getSize().y));
+	this->background.setSize(Vector2f(static_cast<float>(this->window->getSize().x), static_cast<float>(this->window->getSize().y)));
 	if (this->texture.loadFromFile("Assets/Image/Background/mainmenubackground.jpg")) {
 		cout << "Main menu: background loaded!" << endl;
 		this->background.setTexture(&this->texture);
